Send notifications for all visitors in the session
